+++ components/camel-elasticsearch-rest/src/test/java/org/apache/camel/component/elasticsearch/integration/ElasticsearchPingIT.java
@@ -14,14 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.elasticsearch;
+package org.apache.camel.component.elasticsearch.integration;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class ElasticsearchPingTest extends ElasticsearchBaseTest {
+public class ElasticsearchPingIT extends ElasticsearchTestSupport {
 
     @Test
     public void testPing() throws Exception {
